fix(database): enforce auto-migration policy and update schema management
- Updated AGENTS.md to clarify that manual migrations are banned and only auto-migrations are allowed.
- Introduced a new script (fix_schemas.ps1) to automate schema adjustments for Room database files.
- Modified gradle.properties to disable configuration cache for improved build reliability.
- Enhanced detekt.yml with custom rules to prevent manual migrations.
- Updated multiple JSON schema files to ensure default values are set for new columns.
- Removed legacy migration classes and updated database initialization to exclusively use auto-migrations.
- Added DATABASE_MIGRATIONS.md to document migration strategies and guidelines.
