+++ data/src/main/java/gr/eduinvoice/data/model/GroupStudentCrossRef.kt
@@ -1,5 +1,6 @@
 package gr.eduinvoice.data.model
 
+import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.Index
 import kotlinx.serialization.Serializable
@@ -14,5 +15,6 @@ import gr.eduinvoice.data.database.DatabaseConstants
 data class GroupStudentCrossRef(
     val groupId: Long,
     val studentId: Long,
+    @ColumnInfo(defaultValue = "0")
     val ownerId: Long = 0
 )
