+++ data/src/main/java/gr/eduinvoice/data/model/RescheduleMasterLessonLink.kt
@@ -1,14 +1,15 @@
 package gr.eduinvoice.data.model
 
 import androidx.room.Entity
-import androidx.room.Index
+import androidx.room.PrimaryKey
+import kotlinx.serialization.Serializable
 
+@Serializable
 @Entity(
-	tableName = "reschedule_master_lessons",
-	primaryKeys = ["masterId", "lessonId"],
-	indices = [Index(value = ["masterId"]), Index(value = ["lessonId"])]
+    tableName = "reschedule_master_lessons",
+    primaryKeys = ["masterId", "lessonId"]
 )
 data class RescheduleMasterLessonLink(
-	val masterId: Long,
-	val lessonId: Long
+    val masterId: Long,
+    val lessonId: Long
 )
