+++ data/src/main/java/gr/eduinvoice/data/model/Student.kt
@@ -5,12 +5,14 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import kotlinx.serialization.Serializable
+import gr.eduinvoice.data.model.RateTypes
 
 @Serializable
 @Entity(tableName = "students")
 data class Student(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
+    @ColumnInfo(defaultValue = "0")
     val ownerId: Long = 0,
     val name: String,
     @ColumnInfo(defaultValue = "''")
