+++ data/src/main/java/gr/eduinvoice/data/model/StudentGroup.kt
@@ -1,5 +1,6 @@
 package gr.eduinvoice.data.model
 
+import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import kotlinx.serialization.Serializable
@@ -10,10 +11,17 @@ import gr.eduinvoice.data.database.DatabaseConstants
 data class StudentGroup(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0,
+    @ColumnInfo(defaultValue = "0")
     val ownerId: Long = 0,
     val name: String,
+    @ColumnInfo(defaultValue = "''")
     val className: String = "",
+    @ColumnInfo(defaultValue = "0.0")
     val rate: Double = 0.0,
-    val rateType: String = RateTypes.HOURLY,
-    val isActive: Boolean = true
+    @ColumnInfo(defaultValue = "'hourly'")
+    val rateType: String = "hourly",
+    @ColumnInfo(defaultValue = "1")
+    val isActive: Boolean = true,
+    @ColumnInfo(defaultValue = "0")
+    val lastModified: Long = System.currentTimeMillis()
 )
