+++ gradle.properties
@@ -3,11 +3,11 @@
 ###############################################################################
 
 # Gradle performance optimizations
-org.gradle.caching=true
+org.gradle.caching=false
 org.gradle.parallel=true
 org.gradle.configureondemand=true
 org.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=1g -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
-org.gradle.unsafe.configuration-cache=true
+org.gradle.unsafe.configuration-cache=false
 
 # Kotlin optimizations
 kotlin.incremental=true
